api_tests/requirements.txt
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
PyJWT==2.8.0
Werkzeug==3.0.1
requests==2.31.0
